requirements.txt
astroid==2.4.2
azure-functions==1.5.0
azure-functions-durable>=1.0.0b6
azure-storage-blob==1.5.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
colorama==0.4.4
decorator==4.4.2
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.3
isodate==0.6.0
isort==5.7.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
moviepy==1.0.3
msrest==0.6.19
numpy==1.19.5
oauthlib==3.1.0
Pillow==8.1.0
proglog==0.1.9
pycparser==2.20
pylint==2.6.0
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
toml==0.10.2
tqdm==4.56.0
typed-ast==1.4.2
urllib3==1.26.2
wrapt==1.12.1
